+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Windows network breadcrumbs split the server and share, so you can click the server to browse its shares.
+
 ## [0.8.0] - 2026-06-01
 
 ### Added